+++ src/layouts/index.css
@@ -764,8 +764,36 @@ ul {
 .game {
   display: flex;
   flex-direction: row;
+  text-align: center;
 }
 
 .game-info {
   margin-left: 20px;
+}
+
+.styledButton {
+  margin: auto;
+  border: 20px;
+  line-height: 2.5;
+  font-size: 1rem;
+  text-align: center;
+  color: #fff;
+  width: 100px;
+  text-shadow: 1px 1px 1px #000;
+  border-radius: 10px;
+  background-color: rgb(93, 64, 221);
+  background-image: linear-gradient(to top left, rgba(0, 0, 0, .2), rgba(0, 0, 0, .2) 30%, rgba(0, 0, 0, 0));
+  box-shadow: inset 2px 2px 3px rgba(255, 255, 255, .6), inset -2px -2px 3px rgba(0, 0, 0, .6);
+}
+
+.styledButton:hover {
+  background-color: rgba(93, 64, 221, 1);
+}
+
+.styledButton:active {
+  box-shadow: inset -2px -2px 3px rgba(255, 255, 255, .6), inset 2px 2px 3px rgba(0, 0, 0, .6);
+}
+
+.separator {
+  height: 10px;
 }
