+++ src/pages/cellState.tsx
@@ -1,0 +1,5 @@
+export enum state { DEAD, ALIVE }
+export const style = {
+    [state.DEAD]: 'dead',
+    [state.ALIVE]: 'alive',
+};
